pyproject.toml
[project]
name = "rag-study-assistant"
version = "0.1.0"
description = "Local-first RAG study assistant skeleton"
authors = [{ name = "rag-study-assistant" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "streamlit>=1.29",
  "pydantic>=2.6",
  "PyYAML>=6.0",
  "python-dotenv>=1.0",
  "rich>=13.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.5.0",
]

[project.urls]
Homepage = "https://example.com"

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
